Revert "Revert " [clang] Refactor to remove clangDriver dependency from clang…"
This reverts commit dea330b38d9c18b68219abdb52baaa72c9f1103d.
