+++ clang/lib/CrossTU/CMakeLists.txt
@@ -9,6 +9,7 @@ add_clang_library(clangCrossTU
   LINK_LIBS
   clangAST
   clangBasic
+  clangDriver
   clangFrontend
   clangIndex
   )
