+++ clang/lib/Driver/CMakeLists.txt
@@ -17,6 +17,8 @@ endif()
 add_clang_library(clangDriver
   Action.cpp
   Compilation.cpp
+  CreateASTUnitFromArgs.cpp
+  CreateInvocationFromArgs.cpp
   Distro.cpp
   Driver.cpp
   Job.cpp
@@ -96,6 +98,8 @@ add_clang_library(clangDriver
 
   LINK_LIBS
   clangBasic
+  clangFrontend
+  clangSerialization
   clangLex
   clangOptions
   ${system_libs}
