+++ clang/tools/c-index-test/CMakeLists.txt
@@ -27,6 +27,7 @@ else()
     libclang
     clangAST
     clangBasic
+    clangDriver
     clangFrontend
     clangIndex
     clangSerialization
