+++ clang/tools/libclang/CMakeLists.txt
@@ -65,6 +65,7 @@ set(LIBS
   clangFrontend
   clangIndex
   clangLex
+  clangOptions
   clangRewrite
   clangSema
   clangSerialization
